Tiny Resort Tools Release 0.8 (#12)
# Description
TR Tools 0.8.0 release. 

## Developers
### New Features
- Used Newtonsoft.Json to add the ability to read in an array of
quickitems (current the file needs to be named .qitems).
  - Load `qitems` first and then load single `qitem`. 
- Added CompareAPIVersion method that takes in a string of a version,
formatted as "X.X.X".
- Returns -1 if older version, 0 if exact version, and 1 if newer
version
- Added `.qversion` files that let the developer (or user) add the
nexusID, current version, and mod name to use for the mod update
checker.

### Improvements
- Updated chat commands to only lowercase the first two parameters
(trigger and command) and leave the arguments alone for the mod
developer to use in any way they desire.
- Removed the albedo coloring on all quickitem. 
- Added new event `sceneSetupScript` for developers to subscribe to for
anything their mod needs to re-do on a reload to the main menu. We use
this internally for button creation.

## Users

### Improvements
- Added two config options to filter the mod update checker
  - Show/Hide up-to-date mods
  - Show/Hide unknown nexus ID mods

### Bug Fixes
- Fixed bug where clothing and signs were not cleared away on reload of
the game if the custom items were removed from the game.
- Fixed bug where custom icons were not showing up in the correct size
on item signs.
- Fixed bug where when the user quit to the main menu the MODS button
would be missing.
- Fixed bug where some mods with versions not following the Semantic
Version format would not load in the Mod Update Checker window.

## Internal

### New Features
- Added the ability to embed DLLs into the project. This allowed us to
add Newtonsoft.Json.

### Improvements
- Added check to make sure a nexusID + uniqueID combo is unused. 
- Added a line to the TRButton.Copy that ensures the old button actions
are destroyed. This was added due to the weird bug of the game not
removing listeners previously.
- Commented out unused code, added xml information, and disabled the
warning for quickitems script.

### Bug Fixes
- Fixed bug where clothing and bugs on top of other furniture were
undetected due to the custom item being in the status and not the tile
object ID.

Co-authored-by: John Flaherty <slowcircuits@gmail.com>